# Add support for generics