--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,1 @@
 .idea
-_example