--- a/porcupine.go
+++ b/porcupine.go
@@ -3,7 +3,7 @@ package porcupine
 import "time"
 
 // CheckOperations checks whether a history is linearizable.
-func CheckOperations[S State[S]](model Model[S], history []Operation) bool {
+func CheckOperations[S State[S], I any, O any](model Model[S, I, O], history []Operation) bool {
 	res, _ := checkOperations(model, history, false, 0)
 	return res == Ok
 }
@@ -12,7 +12,7 @@ func CheckOperations[S State[S]](model Model[S], history []Operation) bool {
 // timeout.
 //
 // A timeout of 0 is interpreted as an unlimited timeout.
-func CheckOperationsTimeout[S State[S]](model Model[S], history []Operation, timeout time.Duration) CheckResult {
+func CheckOperationsTimeout[S State[S], I any, O any](model Model[S, I, O], history []Operation, timeout time.Duration) CheckResult {
 	res, _ := checkOperations(model, history, false, timeout)
 	return res
 }
@@ -21,12 +21,12 @@ func CheckOperationsTimeout[S State[S]](model Model[S], history []Operation, tim
 // computing data that can be used to visualize the history and linearization.
 //
 // The returned linearizationInfo can be used with [Visualize].
-func CheckOperationsVerbose[S State[S]](model Model[S], history []Operation, timeout time.Duration) (CheckResult, linearizationInfo) {
+func CheckOperationsVerbose[S State[S], I any, O any](model Model[S, I, O], history []Operation, timeout time.Duration) (CheckResult, linearizationInfo) {
 	return checkOperations(model, history, true, timeout)
 }
 
 // CheckEvents checks whether a history is linearizable.
-func CheckEvents[S State[S]](model Model[S], history []Event) bool {
+func CheckEvents[S State[S], I any, O any](model Model[S, I, O], history []Event) bool {
 	res, _ := checkEvents(model, history, false, 0)
 	return res == Ok
 }
@@ -34,7 +34,7 @@ func CheckEvents[S State[S]](model Model[S], history []Event) bool {
 // CheckEventsTimeout checks whether a history is linearizable, with a timeout.
 //
 // A timeout of 0 is interpreted as an unlimited timeout.
-func CheckEventsTimeout[S State[S]](model Model[S], history []Event, timeout time.Duration) CheckResult {
+func CheckEventsTimeout[S State[S], I any, O any](model Model[S, I, O], history []Event, timeout time.Duration) CheckResult {
 	res, _ := checkEvents(model, history, false, timeout)
 	return res
 }
@@ -43,6 +43,6 @@ func CheckEventsTimeout[S State[S]](model Model[S], history []Event, timeout tim
 // data that can be used to visualize the history and linearization.
 //
 // The returned linearizationInfo can be used with [Visualize].
-func CheckEventsVerbose[S State[S]](model Model[S], history []Event, timeout time.Duration) (CheckResult, linearizationInfo) {
+func CheckEventsVerbose[S State[S], I any, O any](model Model[S, I, O], history []Event, timeout time.Duration) (CheckResult, linearizationInfo) {
 	return checkEvents(model, history, true, timeout)
 }