This is a breaking change in interface of Porcupine.

This set of commits does:
* move state to a generic interface, that must implement `Clone`, `String` and `Equals`
* make use of the above methods over the model defined function pointers
* use clone before calling step to avoid issues like in #11 
* use dedicated input/output types for the step function

Note that I'm still looking into refactoring the internal `entry`'s and `node`'value structure to a union type, this proves to be a little more difficult than anticipated - not sure this is currently even possible with two other generic types in Go.