@anishathalye 2022-11-15
This is interesting, thanks for taking time to make the changes and submit a PR.

Btw, I don't think we can do much better for the `entry` and `node` types. Go's type system doesn't support tagged unions, so we're basically manually implementing tagged unions (e.g., the `kind` field of the `entry` is the type tag) there without compile-time checking from the type system.

I see several pros and cons of switching over to using generics.

Pros:

- Avoiding `any` all over the place is nice
- Does help people from missing that `Step()` needs to be purely functional
- Less casting required in user code
- Performance: runtime benefit of avoiding wrapping in interface values in many places

Cons:

- Breaking change to API
- Requires Go 1.18 for generics
- I don't think the types are really necessary here to provide a usable and safe API
- Looks a little messy to have a lot of the functions parameterized by `[S, I, O]`
- Performance: this clones the state everywhere, including for read-only operations, and the user can't opt out of the cloning. `Step()` is on a hot code path.

I was curious about the performance impacts of the change. Looks like there's no statistically significant difference in performance.

Machine stats:

```
goos: linux
goarch: amd64
pkg: github.com/anishathalye/porcupine
cpu: Intel(R) Core(TM) i7-5930K CPU @ 3.50GHz
```

Benchmark results (old = dbc9a3e, new = 4ba1f58).

```
name              old time/op  new time/op  delta
EtcdJepsen000-12   424µs ± 0%   448µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen001-12   505µs ± 0%   507µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen002-12   223ms ± 0%   223ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen003-12   274µs ± 0%   274µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen004-12   155µs ± 0%   149µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen005-12   352µs ± 0%   364µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen006-12   294µs ± 0%   290µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen007-12   185ms ± 0%   182ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen008-12   159µs ± 0%   164µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen009-12   231µs ± 0%   231µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen010-12   158µs ± 0%   157µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen011-12   340µs ± 0%   314µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen012-12   158µs ± 0%   154µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen013-12   140µs ± 0%   137µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen014-12   131µs ± 0%   139µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen015-12   249µs ± 0%   254µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen016-12   127µs ± 0%   131µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen017-12   154µs ± 0%   158µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen018-12   850µs ± 0%   842µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen019-12   488µs ± 0%   495µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen020-12   243µs ± 0%   241µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen021-12   150µs ± 0%   148µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen022-12   146µs ± 0%   145µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen023-12   251µs ± 0%   251µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen024-12   228µs ± 0%   217µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen025-12   213µs ± 0%   218µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen026-12   183µs ± 0%   184µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen027-12   193µs ± 0%   192µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen028-12   223µs ± 0%   216µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen029-12   296µs ± 0%   304µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen030-12   153µs ± 0%   151µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen031-12   159µs ± 0%   162µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen032-12   167µs ± 0%   176µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen033-12   201µs ± 0%   200µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen034-12   170µs ± 0%   177µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen035-12   129µs ± 0%   132µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen036-12   157µs ± 0%   155µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen037-12   329µs ± 0%   327µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen038-12   196µs ± 0%   198µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen039-12   192µs ± 0%   193µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen040-12  21.6ms ± 0%  22.8ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen041-12   133µs ± 0%   134µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen042-12   212µs ± 0%   199µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen043-12   133µs ± 0%   133µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen044-12   480µs ± 0%   462µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen045-12   160µs ± 0%   157µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen046-12   138µs ± 0%   127µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen047-12   253µs ± 0%   234µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen048-12   245µs ± 0%   235µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen049-12   183µs ± 0%   181µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen050-12   120µs ± 0%   119µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen051-12   309µs ± 0%   325µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen052-12   183µs ± 0%   182µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen053-12  4.15ms ± 0%  4.27ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen054-12   156µs ± 0%   165µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen055-12   126µs ± 0%   129µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen056-12   453µs ± 0%   440µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen057-12  93.9ms ± 0%  94.3ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen058-12   160µs ± 0%   158µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen059-12   171µs ± 0%   166µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen060-12   244µs ± 0%   237µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen061-12   572µs ± 0%   529µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen062-12   115µs ± 0%   121µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen063-12   245µs ± 0%   251µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen064-12   138µs ± 0%   135µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen065-12   134µs ± 0%   126µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen066-12   199µs ± 0%   190µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen067-12   203µs ± 0%   200µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen068-12   168µs ± 0%   161µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen069-12   128µs ± 0%   123µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen070-12   133µs ± 0%   130µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen071-12   318µs ± 0%   310µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen072-12   143µs ± 0%   144µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen073-12   843µs ± 0%   859µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen074-12   126µs ± 0%   130µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen075-12  40.1ms ± 0%  39.0ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen076-12   309µs ± 0%   300µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen077-12   140µs ± 0%   141µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen078-12   161µs ± 0%   154µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen079-12   266µs ± 0%   269µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen080-12  14.5ms ± 0%  15.2ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen081-12   146µs ± 0%   142µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen082-12   223µs ± 0%   219µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen083-12   145µs ± 0%   142µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen084-12   213µs ± 0%   212µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen085-12  4.56ms ± 0%  4.71ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen086-12   150µs ± 0%   147µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen087-12   326µs ± 0%   323µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen088-12   166µs ± 0%   167µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen089-12   491µs ± 0%   494µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen090-12   129µs ± 0%   131µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen091-12   128µs ± 0%   127µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen092-12   180µs ± 0%   183µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen093-12   179µs ± 0%   177µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen094-12   140µs ± 0%   137µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen096-12   226µs ± 0%   227µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen097-12  4.62ms ± 0%  4.81ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen098-12  1.68ms ± 0%  1.65ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen099-12  87.6ms ± 0%  89.7ms ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen100-12   201µs ± 0%   201µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen101-12   150µs ± 0%   150µs ± 0%   ~     (p=1.000 n=1+1)
EtcdJepsen102-12   188µs ± 0%   185µs ± 0%   ~     (p=1.000 n=1+1)
```