@anishathalye 2022-11-15
Out of curiosity, I added some new tests and benchmarks based on the key/value store data, and using a model with/without partitioning (ee9b49a57b6e3bc7185369ac814506812db43322, 89abd11ef0922e7b7916741b3a19c6016802b762). Still, no difference between baseline and generics:

```
name                          old time/op  new time/op   delta
Kv1ClientOk-12                 214µs ± 0%    216µs ± 0%   ~     (p=1.000 n=1+1)
Kv1ClientBad-12                106µs ± 0%    107µs ± 0%   ~     (p=1.000 n=1+1)
Kv10ClientsOk-12              1.28ms ± 0%   1.20ms ± 0%   ~     (p=1.000 n=1+1)
Kv10ClientsBad-12              798µs ± 0%    805µs ± 0%   ~     (p=1.000 n=1+1)
Kv50ClientsOk-12               2.73s ± 0%    1.95s ± 0%   ~     (p=1.000 n=1+1)
Kv50ClientsBad-12             13.0ms ± 0%   10.3ms ± 0%   ~     (p=1.000 n=1+1)
KvNoPartition1ClientOk-12      173µs ± 0%    198µs ± 0%   ~     (p=1.000 n=1+1)
KvNoPartition1ClientBad-12    89.2µs ± 0%  102.7µs ± 0%   ~     (p=1.000 n=1+1)
KvNoPartition10ClientsOk-12    90.5s ± 0%    90.8s ± 0%   ~     (p=1.000 n=1+1)
KvNoPartition10ClientsBad-12   63.6s ± 0%    63.9s ± 0%   ~     (p=1.000 n=1+1)
```