@tjungblu 2022-11-16
running the generics without the cloning on the kv benchmarks 

Test | generics ns/op | master ns/op | generics w/o clone ns/op
-- | -- | -- | --
BenchmarkKvNoPartition1ClientBad-12 | 35866 | 33009 | 29800
BenchmarkKv1ClientBad-12 | 41833 | 38036 | 41479
BenchmarkKvNoPartition1ClientOk-12 | 72423 | 63732 | 58706
BenchmarkKv1ClientOk-12 | 80223 | 80575 | 79227
BenchmarkKv10ClientsBad-12 | 306512 | 310970 | 302989
BenchmarkKv10ClientsOk-12 | 488401 | 489597 | 483574
BenchmarkKv50ClientsBad-12 | 9128606 | 11290136 | 8914428
BenchmarkKv50ClientsOk-12 | 1962944648 | 2895008532 | 1968358789
BenchmarkKvNoPartition10ClientsBad-12 | 58994525165 | 56238356959 | 56662378632
BenchmarkKvNoPartition10ClientsOk-12 | 79796681374 | 77089491571 | 77982079434


![image](https://user-images.githubusercontent.com/85170409/202163106-ac924363-2e00-443b-9adc-2a40d10f7fd4.png)